+++ tests/testthat/test-pileup_modbam_cpp.R
@@ -4,19 +4,20 @@ test_that("pileup_modbam_cpp works", {
                               package = "footprintR")
     extractfile <- system.file("extdata", "modkit_extract_rc_6mA_1.tsv.gz",
                                package = "footprintR")
-    
+
     bam4 <- system.file("extdata", "6mA_simple.bam", package = "footprintR")
     bam5 <- system.file("extdata", "6mA_nonPrimary.bam", package = "footprintR")
     bam7 <- system.file("extdata", "6mA_mod-issue.bam", package = "footprintR")
     bam8 <- system.file("extdata", "6mA_too-many-mods.bam", package = "footprintR")
-    
+    bam9 <- system.file("extdata", "6mA_two-mods-on-same-base.bam", package = "footprintR")
+
     ## invalid arguments -------------------------------------------------------
     # ... non-existing bam file
     expect_error(pileup_modbam_cpp(inname_str = "error", regions = "chr1",
                                    modbase = "a", n_threads = 2, level = "read",
                                    mod_prob_thresh = 0.5, verbose = FALSE),
                  "Could not open input file")
-    
+
     # ... no bam index
     tmpbam <- tempfile(fileext = ".bam")
     expect_true(file.copy(from = modbamfile, to = tmpbam))
@@ -30,12 +31,12 @@ test_that("pileup_modbam_cpp works", {
     expect_error(pileup_modbam_cpp(inname_str = bam4, regions = "chr2",
                                    modbase = "a"),
                  "Failed to get bam iterator")
-    
+
     # ... too many modifications on a single base
     expect_error(pileup_modbam_cpp(inname_str = bam8, regions = "chr1",
                                    modbase = "a", verbose = FALSE),
                  "More modifications than footprintR")
-    
+
     ## expected results --------------------------------------------------------
     # ... run pileup_modbam_cpp
     df <- read.delim(extractfile)
@@ -47,35 +48,35 @@ test_that("pileup_modbam_cpp works", {
                                   n_threads = 2,
                                   verbose = TRUE)
     })
-    res2 <- pileup_modbam_cpp(inname_str = modbamfile, 
-                              regions = "chr1:", 
+    res2 <- pileup_modbam_cpp(inname_str = modbamfile,
+                              regions = "chr1:",
                               modbase = "a", level = "read",
-                              n_threads = 1, 
+                              n_threads = 1,
                               verbose = FALSE)
-    res3 <- pileup_modbam_cpp(inname_str = modbamfile, 
-                              regions = c("chr1", "chr2"), 
+    res3 <- pileup_modbam_cpp(inname_str = modbamfile,
+                              regions = c("chr1", "chr2"),
                               modbase = "m", level = "read",
-                              n_threads = 1, 
+                              n_threads = 1,
                               verbose = FALSE)
     res4 <- pileup_modbam_cpp(inname_str = bam4,
                               regions = "chr1",
                               modbase = "a", level = "read",
                               n_threads = 1,
                               verbose = FALSE)
-    res5 <- pileup_modbam_cpp(inname_str = bam5, 
-                              regions = "chr1", 
+    res5 <- pileup_modbam_cpp(inname_str = bam5,
+                              regions = "chr1",
                               modbase = "a", level = "read",
-                              n_threads = 1, 
+                              n_threads = 1,
                               verbose = FALSE)
-    res6a <- pileup_modbam_cpp(inname_str = modbamfile, 
-                               regions = "chr1:6941000-6941001", 
+    res6a <- pileup_modbam_cpp(inname_str = modbamfile,
+                               regions = "chr1:6941000-6941001",
                                modbase = "a", level = "read",
-                               n_threads = 1, 
+                               n_threads = 1,
                                verbose = FALSE)
-    res6b <- pileup_modbam_cpp(inname_str = modbamfile, 
-                               regions = c("chr1:6941000-6941001", "chr1:6928000-6928001"), 
+    res6b <- pileup_modbam_cpp(inname_str = modbamfile,
+                               regions = c("chr1:6941000-6941001", "chr1:6928000-6928001"),
                                modbase = "a", level = "read",
-                               n_threads = 1, 
+                               n_threads = 1,
                                verbose = FALSE)
     aln6a <- Rsamtools::scanBam(file = modbamfile,
                                 param = Rsamtools::ScanBamParam(
@@ -87,7 +88,11 @@ test_that("pileup_modbam_cpp works", {
                                     what = "qname",
                                     which = GRanges(c("chr1:6941000-6941001", "chr1:6928000-6928001"))
                                 ))
-    
+    res9h <- pileup_modbam_cpp(inname_str = bam9, regions = "chr1", modbase = "h",
+                               level = "read", n_threads = 1, verbose = FALSE)
+    res9m <- pileup_modbam_cpp(inname_str = bam9, regions = "chr1", modbase = "m",
+                               level = "read", n_threads = 1, verbose = FALSE)
+
     # ... results structure
     expect_type(res1, "list")
     expect_type(res2, "list")
@@ -96,7 +101,9 @@ test_that("pileup_modbam_cpp works", {
     expect_type(res5, "list")
     expect_type(res6a, "list")
     expect_type(res6b, "list")
-    
+    expect_type(res9h, "list")
+    expect_type(res9m, "list")
+
     expected_names <- c(
         "chrom", "ref_position", "ref_mod_strand", "mod_prob", "read_id", "read_df")
     expect_named(res1, expected_names)
@@ -106,7 +113,9 @@ test_that("pileup_modbam_cpp works", {
     expect_named(res5, expected_names)
     expect_named(res6a, expected_names)
     expect_named(res6b, expected_names)
-    
+    expect_named(res9h, expected_names)
+    expect_named(res9m, expected_names)
+
     expected_types <- c(
         "character", "integer", "character", "double",
         "character", "list")
@@ -118,8 +127,10 @@ test_that("pileup_modbam_cpp works", {
         expect_type(res5[[expected_names[i]]], expected_types[i])
         expect_type(res6a[[expected_names[i]]], expected_types[i])
         expect_type(res6b[[expected_names[i]]], expected_types[i])
+        expect_type(res9h[[expected_names[i]]], expected_types[i])
+        expect_type(res9m[[expected_names[i]]], expected_types[i])
     }
-    
+
     expect_s3_class(res1[["read_df"]], "data.frame")
     expect_s3_class(res2[["read_df"]], "data.frame")
     expect_s3_class(res3[["read_df"]], "data.frame")
@@ -127,7 +138,9 @@ test_that("pileup_modbam_cpp works", {
     expect_s3_class(res5[["read_df"]], "data.frame")
     expect_s3_class(res6a[["read_df"]], "data.frame")
     expect_s3_class(res6b[["read_df"]], "data.frame")
-    
+    expect_s3_class(res9h[["read_df"]], "data.frame")
+    expect_s3_class(res9m[["read_df"]], "data.frame")
+
     expected_df_colnames <- c("read_id", "qscore", "read_length", "aligned_length", "variant_label")
     expect_named(res1$read_df, expected_df_colnames)
     expect_named(res2$read_df, expected_df_colnames)
@@ -136,7 +149,9 @@ test_that("pileup_modbam_cpp works", {
     expect_named(res5$read_df, expected_df_colnames)
     expect_named(res6a$read_df, expected_df_colnames)
     expect_named(res6b$read_df, expected_df_colnames)
-    
+    expect_named(res9h$read_df, expected_df_colnames)
+    expect_named(res9m$read_df, expected_df_colnames)
+
     # ... content res1
     expect_identical(res1$read_df$read_id,
                      c("233e48a7-f379-4dcf-9270-958231125563",
@@ -159,7 +174,7 @@ test_that("pileup_modbam_cpp works", {
     expect_identical(sum(!is.na(i1)), 11183L)
     expect_identical(res1$ref_position[!is.na(i1)],
                      df$ref_position[i1[!is.na(i1)]])
-    
+
     # ... content res2
     expect_identical(res2$read_df$read_id,
                      c("233e48a7-f379-4dcf-9270-958231125563", "d52a5f6a-a60a-4f85-913e-eada84bfbfb9",
@@ -195,41 +210,41 @@ test_that("pileup_modbam_cpp works", {
     expect_true(all(
         res2$call_code[!is.na(i2)] == df$call_code[i2[!is.na(i2)]] |
             res2$mod_prob[!is.na(i2)] < 0.5))
-    
+
     # ... content res3
     for (nm in setdiff(expected_names, "read_df")) {
         expect_length(res3[[nm]], 0L)
     }
-    ## This will still contain all reads - will be filtered out in the 
+    ## This will still contain all reads - will be filtered out in the
     ## readModBam R wrapper
     expect_identical(nrow(res3$read_df), 10L)
-    
+
     # ... content of res4
     expect_equal(res4, list(
         chrom = rep("chr1", 8),
-        ref_position = c(6940001L, 6940004L, 6940008L, 6940010L, 6940012L, 
+        ref_position = c(6940001L, 6940004L, 6940008L, 6940010L, 6940012L,
                          6940015L, 6940017L, 6940019L),
         ref_mod_strand = c("+", "-", "+", "-", "+", "+", "-", "+"),
-        mod_prob = c(0.134765625, 0.318359375, 0.380859375, -1, -1, 
+        mod_prob = c(0.134765625, 0.318359375, 0.380859375, -1, -1,
                      0.724609375, -1, 0.998046875),
-        read_id = c("artificial-read-1", "artificial-read-2", "artificial-read-1", 
-                    "artificial-read-2", "artificial-read-1", "artificial-read-1", 
+        read_id = c("artificial-read-1", "artificial-read-2", "artificial-read-1",
+                    "artificial-read-2", "artificial-read-1", "artificial-read-1",
                     "artificial-read-2", "artificial-read-1"),
         read_df = data.frame(read_id = c("artificial-read-1", "artificial-read-2"),
                              qscore = c(13.4761904761905, 13.24),
                              read_length = c(21L, 25L),
                              aligned_length = c(19L, 23L),
                              variant_label = rep(NA_character_, 2L))))
-    
+
     # ... content of res5
     expect_identical(res5, list(
-        chrom = character(0), ref_position = integer(0), 
+        chrom = character(0), ref_position = integer(0),
         ref_mod_strand = character(0), mod_prob = numeric(0),
-        read_id = character(0), 
+        read_id = character(0),
         read_df = data.frame(read_id = character(0), qscore = numeric(0),
                              read_length = integer(0), aligned_length = integer(0),
                              variant_label = character(0))))
-    
+
     # ... content of res6a and res6b (res6a should be a subset of res6b)
     # ... ... check ground truth
     expect_identical(names(aln6a), names(aln6b)[1])
@@ -248,6 +263,12 @@ test_that("pileup_modbam_cpp works", {
     expect_identical(res6a$mod_prob, res6b$mod_prob[idx])
     expect_equal(res6a$read_df, res6b$read_df[2, , drop = FALSE],
                  ignore_attr = TRUE)
+
+    # ... content of res9h and res9m
+    expect_identical(res9h[!names(res9h) %in% c("call_code", "mod_prob")],
+                     res9m[!names(res9m) %in% c("call_code", "mod_prob")])
+    expect_equal(res9h$mod_prob, (c(25,10) + 0.5) / 256)
+    expect_equal(res9m$mod_prob, (c(230,245) + 0.5) / 256)
 })
 
 test_that("pileup_modbam_cpp works by comparing to read_modbam_cpp", {
@@ -260,10 +281,10 @@ test_that("pileup_modbam_cpp works by comparing to read_modbam_cpp", {
                                 variantRefNames = character(0),
                                 variantRefPositions = integer(0),
                                 n_threads = 1, verbose = FALSE)
-        tmp <- tmp0[c("chrom", "ref_position", "ref_mod_strand", "mod_prob", 
+        tmp <- tmp0[c("chrom", "ref_position", "ref_mod_strand", "mod_prob",
                       "read_id")]
         if (level == "summary") {
-            tmp <- tmp |> 
+            tmp <- tmp |>
                 as.data.frame() |>
                 dplyr::group_by(chrom, ref_position, ref_mod_strand) |>
                 dplyr::summarise(Nmod = sum(mod_prob >= mod_prob_thresh),
@@ -271,8 +292,8 @@ test_that("pileup_modbam_cpp works by comparing to read_modbam_cpp", {
                                  .groups = "drop") |>
                 dplyr::arrange(ref_position, ref_mod_strand)
         } else {
-            tmp <- tmp |> 
-                as.data.frame() |> 
+            tmp <- tmp |>
+                as.data.frame() |>
                 dplyr::arrange(ref_position, read_id, dplyr::desc(ref_mod_strand))
         }
         list(lst = as.list(tmp), df = tmp0$read_df)
@@ -296,14 +317,14 @@ test_that("pileup_modbam_cpp works by comparing to read_modbam_cpp", {
     expect_named(res, c("chrom", "ref_position", "ref_mod_strand",
                         "Nmod", "Nvalid"))
     expect_identical(res0$lst, res)
-    
+
     # reading all alignments in a bam file (read)
     # ... expected results
     modbamfile <- system.file("extdata", "6mA_1_10reads.bam", package = "footprintR")
     reg <- "."
     thresh <- 0.7
     res0 <- get_expected_result(modbamfile, reg, "a", thresh, level = "read")
-    
+
     # ... compare to pileup_modbam_cpp return value
     suppressMessages({
         res <- pileup_modbam_cpp(inname_str = modbamfile, regions = reg,
@@ -317,9 +338,9 @@ test_that("pileup_modbam_cpp works by comparing to read_modbam_cpp", {
     expect_identical(res0$df$read_id, res$read_df$read_id)
     expect_identical(names(res0$lst), names(res[1:5]))
     expect_identical(lengths(res0$lst), lengths(res[1:5]))
-    res <- res[1:5] |> 
-        as.data.frame() |> 
-        dplyr::arrange(ref_position, read_id, dplyr::desc(ref_mod_strand)) |> 
+    res <- res[1:5] |>
+        as.data.frame() |>
+        dplyr::arrange(ref_position, read_id, dplyr::desc(ref_mod_strand)) |>
         as.list()
     expect_identical(res0$lst, res)
 
@@ -339,14 +360,14 @@ test_that("pileup_modbam_cpp works by comparing to read_modbam_cpp", {
     expect_named(res, c("chrom", "ref_position", "ref_mod_strand",
                         "Nmod", "Nvalid"))
     expect_identical(res0$lst, res)
-    
+
     # reading alignments overlapping a region (read)
     # ... expected results
     modbamfile <- system.file("extdata", "6mA_2_10reads.bam", package = "footprintR")
     reg <- "chr1:6935830-6935830"
     thresh <- 0.3
     res0 <- get_expected_result(modbamfile, reg, "a", thresh, "read")
-    
+
     # ... compare to pileup_modbam_cpp return value
     res <- pileup_modbam_cpp(inname_str = modbamfile, regions = reg,
                              modbase = "a", mod_prob_thresh = thresh,
@@ -356,9 +377,9 @@ test_that("pileup_modbam_cpp works by comparing to read_modbam_cpp", {
     expect_named(res, c("chrom", "ref_position", "ref_mod_strand",
                         "mod_prob", "read_id", "read_df"))
     expect_identical(res0$df$read_id, res$read_df$read_id)
-    res <- res[1:5] |> 
-        as.data.frame() |> 
-        dplyr::arrange(ref_position, read_id, dplyr::desc(ref_mod_strand)) |> 
+    res <- res[1:5] |>
+        as.data.frame() |>
+        dplyr::arrange(ref_position, read_id, dplyr::desc(ref_mod_strand)) |>
         as.list()
     expect_identical(res0$lst, res)
 })