+++ tests/testthat/test-scanGenome.R
@@ -1,3 +1,9 @@
+test_that(".calcDirChiSqP works", {
+    expect_equal(.calcDirChiSqP(0, 0, 0, 0), NaN)
+    expect_equal(.calcDirChiSqP(15, 20, 125, 107), -0.6643047)
+    expect_error(.calcDirChiSqP(15, 20, 7, 12), "must be nonnegative")
+})
+
 test_that("genome scanning works (helper functions)", {
     ## example data
     modbamfiles <- system.file("extdata",
@@ -42,6 +48,63 @@ test_that("genome scanning works (helper functions)", {
     expect_equal(start(rg), c(1, 13, 25, 37, 49, 61, 73, 85, 97))
     expect_equal(width(rg), rep(12, 9))
 
+    ## sumNmodNvalid
+    rg <- .tileChromosome(tileSize = 1e6, windowSize = 1e6,
+                          windowStep = 1e6, chromName = "chr1",
+                          chromLength = 70e6)
+    se1 <- se0
+    SummarizedExperiment::assayNames(se1) <- c("a", "b", "c")
+    expect_error(sumNmodNvalid(se1, rg), ".se. must contain assays")
+    res <- sumNmodNvalid(se0, rg)
+    expect_identical(SummarizedExperiment::assay(res, "Nmod")[1, ],
+                     colSums(SummarizedExperiment::assay(se0, "Nmod")))
+    expect_identical(SummarizedExperiment::assay(res, "Nvalid")[1, ],
+                     colSums(SummarizedExperiment::assay(se0, "Nvalid")))
+
+    ## phasingScoreFourier
+    rg <- .tileChromosome(tileSize = 1e6, windowSize = 1e6,
+                          windowStep = 1e6, chromName = "chr1",
+                          chromLength = 70e6)
+    se1 <- se0
+    SummarizedExperiment::assayNames(se1) <- c("a", "b", "c")
+    expect_error(phasingScoreFourier(se1, rg), ".se. must contain assays")
+    expect_error(phasingScoreFourier(
+        se0, GenomicRanges::GRanges("chr1", IRanges::IRanges(start = 1:2, width = 3:4))),
+        "have the same width")
+    expect_error(phasingScoreFourier(
+        se0, GenomicRanges::GRanges("chr1", IRanges::IRanges(start = 1:2, width = 6)),
+        numCoef = 5),
+        "is not divisible")
+    expect_error(phasingScoreFourier(se0, rg[c(1, 3, 4)]),
+                 "need to be regularly spaced")
+    windowgr <- GenomicRanges::GRanges(
+        seqnames = "chr1",
+        ranges = IRanges::IRanges(start = seq(0, 11) * 183 + 6930001,
+                                  width = 4 * 183))
+    res0 <- phasingScoreFourier(se = se0[numeric(0), ], gr = windowgr, numCoef = 5)
+    res1 <- phasingScoreFourier(se = IRanges::subsetByOverlaps(se0, windowgr),
+                                gr = windowgr, numCoef = 5)
+    expect_identical(dim(res0), c(0L, BiocGenerics::ncol(se0)))
+    expect_identical(dim(res1), c(length(windowgr), BiocGenerics::ncol(se0)))
+    expect_identical(SummarizedExperiment::assayNames(res0),
+                     c("phasingScoreAbs", "phasingScoreRel"))
+    expect_identical(SummarizedExperiment::assayNames(res1),
+                     c("phasingScoreAbs", "phasingScoreRel"))
+    expect_identical(SummarizedExperiment::rowRanges(res1), windowgr)
+    ass1 <- SummarizedExperiment::assay(res1, "phasingScoreAbs")
+    ass2 <- SummarizedExperiment::assay(res1, "phasingScoreRel")
+    expect_type(ass1, "double")
+    expect_identical(ass1[, 1], ass1[, 2])
+    expect_identical(ass1[, 3], ass1[, 4])
+    expect_true(all(ass1 >= 0))
+    expect_type(ass2, "double")
+    expect_identical(ass2[, 1], ass2[, 2])
+    expect_identical(ass2[, 3], ass2[, 4])
+    expect_true(all(ass2 >= 0 & ass2 <= 1.0))
+    corr1 <- cor(ass1[, c(1, 3)], ass2[, c(1, 3)])
+    expect_equal(which.max(corr1[, 1]), 1L, ignore_attr = TRUE)
+    expect_equal(which.max(corr1[, 2]), 2L, ignore_attr = TRUE)
+
     ## quantifyWindowsInRegion
     expect_error(quantifyWindowsInRegion(bamfiles = "error",
                                          region = "chr1:6940000-6955000",
@@ -49,6 +112,11 @@ test_that("genome scanning works (helper functions)", {
     expect_error(quantifyWindowsInRegion(bamfiles = modbamfiles,
                                          region = "error",
                                          modbase = "a"))
+    expect_error(quantifyWindowsInRegion(bamfiles = modbamfiles,
+                                         region = "chr1:6940000-6955000",
+                                         modbase = "a",
+                                         quantFunction = "error"),
+                 ".quantFunction. must be the name of an existing function")
     expect_identical(dim(quantifyWindowsInRegion(bamfiles = modbamfiles,
                                                  region = "chr1:1-1000",
                                                  modbase = "a",
@@ -65,12 +133,12 @@ test_that("genome scanning works (helper functions)", {
     ))
     se1$group <- c("group1", "group1", "group2", "group2")
     expect_s4_class(se1, "RangedSummarizedExperiment")
-    expect_identical(dim(se1), c(134L, 4L))
+    expect_identical(dim(se1), c(136L, 4L))
     expect_true(all(width(SummarizedExperiment::rowRanges(se1)) == 24L))
     expect_identical(SummarizedExperiment::assayNames(se1),
-                     c("Nmod", "Nvalid"))
+                     c("Nmod", "Nvalid", "FracMod"))
     expect_identical(colSums(SummarizedExperiment::assay(se1, "Nvalid")),
-                     c(s1 = 1601, s2 = 1601, s3 = 554, s4 = 554))
+                     c(s1 = 1610, s2 = 1610, s3 = 554, s4 = 554))
     ov <- findOverlaps(query = SummarizedExperiment::rowRanges(se0),
                        subject = SummarizedExperiment::rowRanges(se1))
     Nvalid <- SummarizedExperiment::assay(se0, "Nvalid")
@@ -96,15 +164,15 @@ test_that("genome scanning works (helper functions)", {
                                    sequenceContext = "A",
                                    BPPARAM = BiocParallel::SerialParam())
     expect_s4_class(se2, "RangedSummarizedExperiment")
-    expect_identical(dim(se2), c(134L, 4L))
+    expect_identical(dim(se2), c(136L, 4L))
     expect_true(all(width(SummarizedExperiment::rowRanges(se2)) == 24L))
     i <- GenomicRanges::match(SummarizedExperiment::rowRanges(se2),
                               SummarizedExperiment::rowRanges(se1))
     expect_true(!any(is.na(i)))
     expect_identical(SummarizedExperiment::assayNames(se2),
-                     c("Nmod", "Nvalid"))
+                     c("Nmod", "Nvalid", "FracMod"))
     expect_identical(colSums(SummarizedExperiment::assay(se2, "Nvalid")),
-                     c(s1 = 1581, s2 = 1581, s3 = 540, s4 = 540))
+                     c(s1 = 1590, s2 = 1590, s3 = 540, s4 = 540))
     expect_true(all(SummarizedExperiment::assay(se1, "Nmod")[i,] >= SummarizedExperiment::assay(se2, "Nmod")))
     expect_true(all(SummarizedExperiment::assay(se1, "Nvalid")[i,] >= SummarizedExperiment::assay(se2, "Nvalid")))
     expect_identical(SummarizedExperiment::colData(se1),
@@ -125,14 +193,14 @@ test_that("genome scanning works (helper functions)", {
         gr1 <- getDifferentiallyModifiedWindows(se1, groupCol = "group", verbose = TRUE)
     ))
     expect_s4_class(gr1, "GRanges")
-    expect_length(gr1, 134L)
+    expect_length(gr1, 136L)
     expect_identical(ncol(GenomicRanges::mcols(gr1)), 6L)
     expect_identical(colnames(GenomicRanges::mcols(gr1)),
                      c("logFC", "logCPM", "LR", "PValue", "FDR", "dirNegLog10PValue"))
 
     gr2 <- getDifferentiallyModifiedWindows(se2, groupCol = "group")
     expect_s4_class(gr2, "GRanges")
-    expect_length(gr2, 134L)
+    expect_length(gr2, 136L)
     expect_identical(ncol(GenomicRanges::mcols(gr2)), 6L)
     expect_identical(colnames(GenomicRanges::mcols(gr2)),
                      c("logFC", "logCPM", "LR", "PValue", "FDR", "dirNegLog10PValue"))
@@ -140,26 +208,63 @@ test_that("genome scanning works (helper functions)", {
     expect_true(!any(is.na(i)))
     expect_true(cor(gr1$logFC[i], gr2$logFC) > 0.98)
 
-    ## fuseWindows
-    expect_error(fuseWindows(x = "error"))
-    expect_error(fuseWindows(x = gr1, scoreCol = "error"))
-    expect_length(fuseWindows(x = gr1[numeric(0)]), 0L)
+    ## getRangesWithAssayValues
+    resL <- list(getRangesWithAssayValues(se0),
+                 getRangesWithAssayValues(se0, "Nmod"),
+                 getRangesWithAssayValues(se0, "FracMod"))
+    tst <- lapply(resL, \(x) expect_s4_class(x, class = "GRanges"))
+    tst <- lapply(resL, \(x) expect_length(x, nrow(se0)))
+    tst <- lapply(resL, \(x) {
+        xx <- x
+        GenomicRanges::mcols(xx) <- NULL
+        expect_identical(xx, SummarizedExperiment::rowRanges(se0))
+    })
+    tmpNmod <- SummarizedExperiment::assay(se0, "Nmod")
+    colnames(tmpNmod) <- paste0("Nmod.", colnames(tmpNmod))
+    expect_identical(as.matrix(GenomicRanges::mcols(resL[[1]])),
+                     tmpNmod)
+    expect_identical(as.matrix(GenomicRanges::mcols(resL[[2]])),
+                     tmpNmod)
+    tmpFracMod <- SummarizedExperiment::assay(se0, "FracMod")
+    colnames(tmpFracMod) <- paste0("FracMod.", colnames(tmpFracMod))
+    expect_identical(as.matrix(GenomicRanges::mcols(resL[[3]])),
+                     tmpFracMod)
 
+    ## processWindowScores
+    expect_error(processWindowScores(x = "error"))
+    expect_error(processWindowScores(x = gr1, scoreCol = "error"))
+    expect_length(processWindowScores(x = gr1[numeric(0)]), 0L)
+
+    # ... pass
+    expect_identical(processWindowScores(x = gr1, scoreCol = "logFC", scoreAction = "pass"), gr1)
+
+    # ... select
+    grs <- processWindowScores(x = gr1, scoreCol = "logFC", scoreAction = "select")
+    expect_identical(gr1$logFC, grs$logFC)
+    expect_identical(colnames(mcols(grs)), "logFC")
+
+    # ... smooth
+    gr1smooth <- processWindowScores(x = gr1, scoreCol = "logFC", scoreAction = "smooth")
+    expect_identical(IRanges::ranges(gr1), IRanges::ranges(gr1smooth))
+    expect_identical(GenomicRanges::mcols(gr1)[, -1], GenomicRanges::mcols(gr1smooth)[, -1])
+    expect_true(cor(GenomicRanges::mcols(gr1)[, 1], GenomicRanges::mcols(gr1smooth)[, 1]) > 0.9)
+
+    # ... smoothFuse
     suppressMessages(expect_message(
-        gr1Fused <- fuseWindows(x = gr1, scoreCol = "logFC", thresh = 5.0, verbose = TRUE)
+        gr1Fused <- processWindowScores(x = gr1, scoreCol = "logFC", thresh = 5.0, verbose = TRUE)
     ))
     expect_s4_class(gr1Fused, "GRanges")
     expect_length(gr1Fused, 5L)
     expect_identical(colnames(GenomicRanges::mcols(gr1Fused)),
                      c("logFCThresh", "numWindowsThresh", "direction", "logFC", "numWindows"))
-    expect_identical(sum(width(gr1Fused)), 768L)
+    expect_identical(sum(width(gr1Fused)), 732L)
 
-    gr2Fused <- fuseWindows(x = gr2, scoreCol = "logFC", thresh = 5.0)
+    gr2Fused <- processWindowScores(x = gr2, scoreCol = "logFC", thresh = 5.0)
     expect_s4_class(gr2Fused, "GRanges")
     expect_length(gr2Fused, 5L)
     expect_identical(colnames(GenomicRanges::mcols(gr2Fused)),
                      c("logFCThresh", "numWindowsThresh", "direction", "logFC", "numWindows"))
-    expect_identical(sum(width(gr2Fused)), 768L)
+    expect_identical(sum(width(gr2Fused)), 732L)
 })
 
 test_that("genome scanning works (wrapper function)", {
@@ -186,4 +291,13 @@ test_that("genome scanning works (wrapper function)", {
                                     modbase = "a", BPPARAM = BiocParallel::SerialParam())
     expect_s4_class(gr, "GRanges")
     expect_length(gr, 38L)
+
+    gr2 <- scanForHighScoringRegions(bamfiles = modbamfiles,
+                                     sampleAnnot = annotdf,
+                                     chromosomeLengths = chrlen,
+                                     scoreCol = c("dirNegLog10PValue", "logFC"),
+                                     modbase = "a", BPPARAM = BiocParallel::SerialParam())
+    expect_s4_class(gr2$dirNegLog10PValue, "GRanges")
+    expect_length(gr2$dirNegLog10PValue, 38L)
+    expect_identical(gr, gr2$dirNegLog10PValue)
 })